splitting appliance to several plugins
