+++ cfme/utils/dockerbot/pytestbase/setup.sh
@@ -260,7 +260,7 @@ set +e
 if [ "$USE_SPROUT" = "yes" ];
 then
     log "invoking complete collectonly with dummy instance before test"
-    gate "collectonly.txt" "py.test --collectonly --dummy-appliance --appliance-version $SPROUT_GROUP --use-provider complete"
+    gate "collectonly.txt" "py.test --collectonly --dummy-appliance --app-version $SPROUT_GROUP --use-provider complete"
 
     run_n_log "miq sprout checkout --populate-yaml --user-key sprout" &
     sleep 5
